+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Unreleased
 
+* Add unix domain socket support for Datadog StatsD metrics
 * Bump minimum rdkafka gem version to 0.15.0
 * Bump minimum Ruby version to 3.0